tweak a little bit with thread

+++ BTL/clientserver.h
@@ -1,14 +1,16 @@
-#include <iostream>
-#include <sstream>
-#include <stdio.h>
-#include <cstring>
-#include <ctime>
-#include <poll.h>
-#include <fstream>
-#include <unordered_map>
+#include <iostream>			// C++
+#include <stdio.h>			// C
+#include <sstream>			// String stream
+#include <fstream>			// File stream
+#include <cstring>			// C String
+#include <ctime>			// Time library
+#include <poll.h>			// Listen on file descriptor events
+#include <unordered_map>	// Map for marker
+#include <sys/mman.h>		// Map file to memory
 #include "udp.pb.h"     	// Protobuf
 #include "network.h"		// Network
 #include "md5.h"			// Hash function
+#include "thread_helper.h"	// Multi thread
 
 class Sockpeer
 {
@@ -19,7 +21,7 @@ private:
 	BTL::ClientInfo* peers;
 	// buffer size, should I let user decide?
 	size_t BUFFSIZE;
-	int dataSize;
+	size_t dataSize;
 	//
 	int localPort;
 	//
@@ -29,18 +31,18 @@ private:
 protected:
 	// Serialize message
 	std::string wrapMessage(BTL::MessageType::Message msgType, google::protobuf::Message* msgData);
-	// Parse message
+	// Parse message -> should parse message by hand
 	// bool parseMessage(BTL::MessageType* internal_MessageType, std::string* BTLMessageString, std::string dataIn);
 	// bool parseMessage(BTL::MessageType* internal_MessageType, google::protobuf::Message* internal_Message, std::string dataIn);
 
 public:
 	// Return true if connected to network, false otherwise
 	bool connected;
-	// 
+	// Return true if this peer is server
 	bool isServer;
 	// Main run loop, read sdk
 	void run();
-	//
+	// Contructor
 	Sockpeer(int localPort, std::string remoteHost, int remotePort, bool isServer);
 	// Destructor
 	~Sockpeer();
